chore: fixing imports small refactor
